requirements/main.txt
gradio==3.36.1
tqdm==4.65.0
numpy==1.23.5
faiss-cpu==1.7.3
fairseq==0.12.2
matplotlib==3.7.1
scipy==1.9.3
librosa==0.9.1
pyworld==0.3.2
soundfile==0.12.1
ffmpeg-python==0.2.0
pydub==0.25.1
soxr==0.3.5
transformers==4.28.1
torchcrepe==0.0.20

tensorboard
tensorboardX
requests